ZitiSSLSocket: misc fixes

+++ ziti/src/samples/java/org/openziti/samples/SampleOkHttp.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.openziti.sample;
+package org.openziti.samples;
 
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
@@ -25,10 +25,9 @@ import org.openziti.ZitiContext;
 import javax.net.ssl.TrustManagerFactory;
 import javax.net.ssl.X509TrustManager;
 import java.net.InetAddress;
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
 import java.security.KeyStore;
 import java.util.Collections;
+import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
 public class SampleOkHttp {
@@ -41,8 +40,24 @@ public class SampleOkHttp {
         try {
             ctx = Ziti.newContext(path, new char[0]);
 
-            // sleep to let ZitiContext initialize
-            Thread.sleep(3000);
+            var active = new CountDownLatch(1);
+            ctx.onStatus(status -> {
+                System.out.println(status);
+                switch (status) {
+                    case ZitiContext.Status.Loading l:
+                        break;
+                    case ZitiContext.Status.Active a:
+                        active.countDown();
+                        break;
+                    default:
+                        System.out.println("Unexpected value: " + status);
+                }
+            });
+
+            if (!active.await(3, TimeUnit.SECONDS)) {
+                System.err.println("failed to load ziti identity");
+                System.exit(1);
+            }
 
             KeyStore ks = KeyStore.getInstance(KeyStore.getDefaultType());
             TrustManagerFactory tmf = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
@@ -71,11 +86,12 @@ public class SampleOkHttp {
             Response resp = clt.newCall(req).execute();
             System.out.println(resp);
             System.out.println(resp.headers());
-            System.out.println(StandardCharsets.UTF_8.decode(ByteBuffer.wrap(resp.body().bytes())));
+            System.out.println(resp.body().string());
         } catch(Exception ex) {
             ex.printStackTrace();
         } finally {
-            ctx.destroy();
+            if (ctx != null)
+                ctx.destroy();
         }
     }
 }
